requirements.txt
Flask==3.0.1
gunicorn==22.0.0
langchain==0.1.17
langchain_openai==0.1.4
weaviate-client==4.4.4